refactor(cuda): decompose CustomStorage helper
